+++ apps/meteor/app/api/server/v1/oauthapps.ts
@@ -1,22 +1,65 @@
+import type { IOAuthApps } from '@rocket.chat/core-typings';
 import { OAuthApps } from '@rocket.chat/models';
-import { isUpdateOAuthAppParams, isOauthAppsGetParams, isOauthAppsAddParams, isDeleteOAuthAppParams } from '@rocket.chat/rest-typings';
+import {
+	ajv,
+	isUpdateOAuthAppParams,
+	isOauthAppsGetParams,
+	isDeleteOAuthAppParams,
+	validateUnauthorizedErrorResponse,
+	validateBadRequestErrorResponse,
+	validateForbiddenErrorResponse,
+} from '@rocket.chat/rest-typings';
 
 import { hasPermissionAsync } from '../../../authorization/server/functions/hasPermission';
 import { apiDeprecationLogger } from '../../../lib/server/lib/deprecationWarningLogger';
 import { addOAuthApp } from '../../../oauth2-server-config/server/admin/functions/addOAuthApp';
 import { deleteOAuthApp } from '../../../oauth2-server-config/server/admin/methods/deleteOAuthApp';
 import { updateOAuthApp } from '../../../oauth2-server-config/server/admin/methods/updateOAuthApp';
+import type { ExtractRoutesFromAPI } from '../ApiClass';
 import { API } from '../api';
 
-API.v1.addRoute(
+const oauthAppsListEndpoints = API.v1.get(
 	'oauth-apps.list',
-	{ authRequired: true, permissionsRequired: ['manage-oauth-apps'] },
 	{
-		async get() {
-			return API.v1.success({
-				oauthApps: await OAuthApps.find().toArray(),
-			});
+		authRequired: true,
+		query: ajv.compile<{ uid?: string }>({
+			type: 'object',
+			properties: {
+				uid: {
+					type: 'string',
+				},
+			},
+			additionalProperties: false,
+		}),
+		permissionsRequired: ['manage-oauth-apps'],
+		response: {
+			400: validateBadRequestErrorResponse,
+			401: validateUnauthorizedErrorResponse,
+			403: validateForbiddenErrorResponse,
+			200: ajv.compile<{ oauthApps: IOAuthApps[] }>({
+				type: 'object',
+				properties: {
+					oauthApps: {
+						type: 'array',
+						items: {
+							$ref: '#/components/schemas/IOAuthApps',
+						},
+					},
+					success: {
+						type: 'boolean',
+						enum: [true],
+					},
+				},
+				required: ['oauthApps', 'success'],
+				additionalProperties: false,
+			}),
 		},
+	},
+
+	async function action() {
+		return API.v1.success({
+			oauthApps: await OAuthApps.find().toArray(),
+		});
 	},
 );
 
@@ -83,18 +126,72 @@ API.v1.addRoute(
 	},
 );
 
-API.v1.addRoute(
+export type OauthAppsAddParams = {
+	name: string;
+	active: boolean;
+	redirectUri: string;
+};
+
+const OauthAppsAddParamsSchema = {
+	type: 'object',
+	properties: {
+		name: {
+			type: 'string',
+		},
+		active: {
+			type: 'boolean',
+		},
+		redirectUri: {
+			type: 'string',
+		},
+	},
+	required: ['name', 'active', 'redirectUri'],
+	additionalProperties: false,
+};
+
+const isOauthAppsAddParams = ajv.compile<OauthAppsAddParams>(OauthAppsAddParamsSchema);
+
+const oauthAppsCreateEndpoints = API.v1.post(
 	'oauth-apps.create',
 	{
 		authRequired: true,
-		validateParams: isOauthAppsAddParams,
+		body: isOauthAppsAddParams,
 		permissionsRequired: ['manage-oauth-apps'],
-	},
-	{
-		async post() {
-			const application = await addOAuthApp(this.bodyParams, this.userId);
-
-			return API.v1.success({ application });
+		response: {
+			400: validateBadRequestErrorResponse,
+			401: validateUnauthorizedErrorResponse,
+			403: validateForbiddenErrorResponse,
+			200: ajv.compile<{ application: IOAuthApps }>({
+				type: 'object',
+				properties: {
+					application: { $ref: '#/components/schemas/IOAuthApps' },
+					success: {
+						type: 'boolean',
+						enum: [true],
+					},
+				},
+				required: ['application', 'success'],
+				additionalProperties: false,
+			}),
 		},
 	},
+
+	async function action() {
+		const application = await addOAuthApp(this.bodyParams, this.userId);
+
+		return API.v1.success({ application });
+	},
 );
+
+type OauthAppsCreateEndpoints = ExtractRoutesFromAPI<typeof oauthAppsCreateEndpoints>;
+
+type OauthAppsListEndpoints = ExtractRoutesFromAPI<typeof oauthAppsListEndpoints>;
+
+export type OAuthAppsEndpoints = OauthAppsCreateEndpoints | OauthAppsListEndpoints;
+
+declare module '@rocket.chat/rest-typings' {
+	// eslint-disable-next-line @typescript-eslint/naming-convention, @typescript-eslint/no-empty-interface
+	interface Endpoints extends OauthAppsCreateEndpoints {}
+	// eslint-disable-next-line @typescript-eslint/naming-convention, @typescript-eslint/no-empty-interface
+	interface Endpoints extends OauthAppsListEndpoints {}
+}