+++ apps/meteor/app/webrtc/client/WebRTCClass.ts
@@ -7,7 +7,7 @@ import { ReactiveVar } from 'meteor/reactive-var';
 
 import { ChromeScreenShare } from './screenShare';
 import { goToRoomById } from '../../../client/lib/utils/goToRoomById';
-import { Subscriptions, Users } from '../../models/client';
+import { Subscriptions, Users } from '../../../client/stores';
 import { settings } from '../../settings/client';
 import { sdk } from '../../utils/client/lib/SDKClient';
 import { t } from '../../utils/lib/i18n';
@@ -820,25 +820,21 @@ class WebRTCClass {
 			return;
 		}
 
-		const user = Users.findOne(data.from);
-		let fromUsername = undefined;
-		if (user?.username) {
-			fromUsername = user.username;
-		}
+		const username = data.from ? Users.state.get(data.from)?.username : undefined;
 		const subscription = Subscriptions.state.find(({ rid }) => rid === data.room);
 
 		let icon;
 		let title;
 		if (data.monitor === true) {
 			icon = 'eye' as const;
-			title = t('WebRTC_monitor_call_from_%s', fromUsername);
+			title = t('WebRTC_monitor_call_from_%s', username);
 		} else if (subscription && subscription.t === 'd') {
 			if (data.media?.video) {
 				icon = 'video' as const;
-				title = t('WebRTC_direct_video_call_from_%s', fromUsername);
+				title = t('WebRTC_direct_video_call_from_%s', username);
 			} else {
 				icon = 'phone' as const;
-				title = t('WebRTC_direct_audio_call_from_%s', fromUsername);
+				title = t('WebRTC_direct_audio_call_from_%s', username);
 			}
 		} else if (data.media?.video) {
 			icon = 'video' as const;